+++ src/main/java/org/betonquest/betonquest/compatibility/protocollib/ProtocolLibIntegrator.java
@@ -39,7 +39,7 @@ public class ProtocolLibIntegrator implements Integrator {
             throw new UnsupportedVersionException(protocolLib, "5.0.0-SNAPSHOT-636");
         }
 
-        plugin.getFeatureRegistries().conversationIO().register("menu", new MenuConvIOFactory(plugin, plugin.getMessageParser(),
+        plugin.getFeatureRegistries().conversationIO().register("menu", new MenuConvIOFactory(plugin, plugin.getTextParser(),
                 plugin.getFontRegistry(), plugin.getPluginConfig(), plugin.getConversationColors()));
         plugin.getFeatureRegistries().interceptor().register("packet", new PacketInterceptorFactory());