+++ src/main/java/org/betonquest/betonquest/feature/QuestCompass.java
@@ -1,6 +1,6 @@
 package org.betonquest.betonquest.feature;
 
-import org.betonquest.betonquest.api.message.Message;
+import org.betonquest.betonquest.api.text.Text;
 import org.betonquest.betonquest.id.ItemID;
 import org.betonquest.betonquest.instruction.variable.Variable;
 import org.bukkit.Location;
@@ -13,5 +13,5 @@ import org.jetbrains.annotations.Nullable;
  * @param location the compass location
  * @param itemID   possible item id, when it should be displayed in the backpack
  */
-public record QuestCompass(Message names, Variable<Location> location, @Nullable ItemID itemID) {
+public record QuestCompass(Text names, Variable<Location> location, @Nullable ItemID itemID) {
 }