+++ src/main/java/org/betonquest/betonquest/quest/event/notify/NotifyEvent.java
@@ -1,10 +1,10 @@
 package org.betonquest.betonquest.quest.event.notify;
 
-import org.betonquest.betonquest.api.message.Message;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
 import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.api.quest.event.PlayerEvent;
 import org.betonquest.betonquest.api.quest.event.online.OnlineEvent;
+import org.betonquest.betonquest.api.text.Text;
 import org.betonquest.betonquest.notify.NotifyIO;
 
 /**
@@ -19,21 +19,21 @@ public class NotifyEvent implements OnlineEvent {
     /**
      * The translations to use.
      */
-    private final Message messages;
+    private final Text text;
 
     /**
      * Creates a new {@link NotifyEvent}.
      *
      * @param notifyIO the {@link NotifyIO} to use
-     * @param messages the messages to use
+     * @param text     the text to use
      */
-    public NotifyEvent(final NotifyIO notifyIO, final Message messages) {
+    public NotifyEvent(final NotifyIO notifyIO, final Text text) {
         this.notifyIO = notifyIO;
-        this.messages = messages;
+        this.text = text;
     }
 
     @Override
     public void execute(final OnlineProfile profile) throws QuestException {
-        notifyIO.sendNotify(messages.asComponent(profile), profile);
+        notifyIO.sendNotify(text.asComponent(profile), profile);
     }
 }